tutorial/data-generator/requirements.txt
confluent-kafka==2.0.2
requests==2.27.1
avro==1.11.1
schedule==1.1.0
svix-ksuid==0.6.0
mysql-connector-python==8.0.33